backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
psycopg2
psycopg2-binary